fixed a severe issue in the ThreadPool implementation.
of course it is possible that they wanted to have it that way,
but the impact on performance was just not acceptable.

the threadpool now has a counter for busy threads - that is for threads
that are currently doing a job. if all threads are busy and another
job is coming in, a new thread will be spawned.

the previous implementation waited for one of the worker threads
to become free, as the result the incoming job was delayed.
this had a severe impact on the webserver performance